Auth feature Completion

+++ routes/postRoutes.js
@@ -4,11 +4,27 @@ const mongoose = require("mongoose");
 const router = express.Router();
 const Post = require('../models/PostModel');
 const Comment = require("../models/CommentModel");
+const User = require('../models/UserModel');
 
 router.get('/', async (req, res) => {
   try{
     let posts = await Post.find().sort({createdAt: -1});
-    res.render('home', {posts});
+
+    let userids = [];
+    for(let post of posts){
+      if (!(userids.includes(post.userId))){
+        userids.push(post.userId);
+      }
+    }
+
+    let postUsers = await User.find({ _id: { $in: userids } });
+    let usersWithPosts = {};
+
+    for(let everyUser of postUsers) {
+      usersWithPosts[`${everyUser._id}`] = everyUser;
+    }
+
+    res.render('home', {posts, user: req.user, usersWithPosts});
   }catch(e) {
     console.error("Something is wrong", e);
   }
@@ -20,7 +36,8 @@ router.get("/home", (req, res) => {
 
 router.post('/posts', async(req, res) => {
   try{
-    let newPost = new Post(req.body);
+    let postObj = {...req.body, userId: req.user._id}
+    let newPost = new Post(postObj);
     await newPost.save();
     res.redirect('/');
   }catch(e) {
@@ -31,10 +48,16 @@ router.post('/posts', async(req, res) => {
 router.get("/editPost/:id", async (req, res) => {
   try {
     let post = await Post.findById(req.params.id);
+    if(post.userId != req.user.id) {
+      throw new Error();
+    }
+    
     console.log(post);
-    res.render('editPost', {post});
+    res.render('editPost', {post, user: req.user});
   }catch(e) {
-    console.error(e)
+    console.error("Something terrible happened here ", e);
+    req.flash("error", "You are NOT authorized to edit this post")
+    res.redirect("/");
   }
 })
 
@@ -65,8 +88,21 @@ router.get("/posts/:id", async (req, res) => {
       }
     ])
     if(singlePost.length > 0) {
+      let userids = [singlePost[0].userId];
+      singlePost[0].comments.forEach( comm => {
+        if(!(userids.includes(comm.userId))){
+          userids.push(comm.userId);
+        }
+      })
+
+      let commUsers = await User.find({ _id: {$in: userids} });
+      let commentsUsers = {};
+      commUsers.forEach( user => {
+        commentsUsers[`${user._id}`] = user;
+      })
+
       singlePost[0].comments = singlePost[0].comments.reverse();
-      res.render("showPost", {post: singlePost[0]});
+      res.render("showPost", {post: singlePost[0], user: req.user, commentsUsers});
     }else {
       throw new Error();
     }
@@ -77,10 +113,19 @@ router.get("/posts/:id", async (req, res) => {
 
 router.delete("/posts/:id", async (req, res) => {
   try {
+    let post = await Post.findById(req.params.id);
+    if(post.userId != req.user.id) {
+      throw new Error();
+    }
+
+    await Comment.deleteMany({postId: req.params.id})
     await Post.findByIdAndDelete(req.params.id);
+    req.flash("success", "Post deleted successfully!")
     res.redirect("/");
   }catch(e) {
     console.error(e);
+    req.flash("error", "You CANNOT delete that post");
+    res.redirect('/');
   }
 })
